moved routing to routes file and imported to index
